print_byte via jmp instruction did not return
When the print_byte function is called via a jmp instruction it is meant as a
tail call. In our case we just printed the byte and continued executing the
code below the jmp instruction. This commit introduces a new virtual
instruction which prints the byte and executes the ret instruction.